Modificaciones en ProductoEndpoints

+++ backend/Novit.Academia/Endpoints/ProductoEndpoints.cs
@@ -1,4 +1,5 @@
 ﻿using Carter;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Novit.Academia.Database;
 using Novit.Academia.Domain;
 using Novit.Academia.Endpoints.DTO;
@@ -21,14 +22,20 @@ public class ProductoEndpoints : ICarterModule
         // GET - Obtener un producto según su id
         app.MapGet("/{idProducto}", (AppDbContext context,int idProducto) =>
         {
-            var producto = context.Productos.FirstOrDefault(p => p.IdProducto == idProducto);
+            var producto = context.Productos.Where(p => p.IdProducto == idProducto).Select(p => p.ConvertToProductoDto());
             return Results.Ok(producto);
         }).WithTags("Producto");
 
         // POST - Crear un producto nuevo
         app.MapPost("/", (AppDbContext context, ProductoDto productoDto) =>
         {
-            Producto producto = productoDto.ConvertToProducto();
+            Producto producto = new() { 
+                Nombre = productoDto.Nombre,
+                Descripcion = productoDto.Descripcion,
+                Precio = productoDto.Precio,
+                Stock = productoDto.Stock,
+                UrlImagen = productoDto.UrlImagen,
+            };
 
             context.Productos.Add(producto);
             context.SaveChanges();
@@ -39,6 +46,9 @@ public class ProductoEndpoints : ICarterModule
         app.MapPut("/{idProducto}", (AppDbContext context,int idProducto, ProductoDto productoDto) =>
         {
             var producto = context.Productos.FirstOrDefault(p => p.IdProducto == idProducto);
+
+            if (producto == null)
+                return Results.BadRequest();
 
             producto.Nombre = productoDto.Nombre;
             producto.Descripcion = productoDto.Descripcion;
@@ -55,6 +65,10 @@ public class ProductoEndpoints : ICarterModule
         app.MapDelete("/{idProducto}", (AppDbContext context ,int idProducto) =>
         {
             var producto = context.Productos.FirstOrDefault(p => p.IdProducto == idProducto);
+
+            if (producto == null)
+                return Results.BadRequest();
+
             context.Productos.Remove(producto);
             context.SaveChanges();
